Read arbitrarily long response (#41)
* Added ability to read arbitrarily long response

* Better method for getting response start; updated doxygen

* Separate out gps and ssl from modem

* Buffer overflow check when getting chunk length

* Correct error in ci_test

* Remove GPS test, add autosampler test
