+++ perfect-cell.cydsn/modem.c
@@ -1,7 +1,7 @@
 /**
  * @file modem.c
  * @brief Implements functions for connecting to the cellular network, and sending/receiving packets.
- * @author Brandon Wong and Matt Bartos
+ * @author Brandon Wong, Matt Bartos and Ivan Mondragon
  * @version TODO
  * @date 2017-06-01
  */
@@ -13,55 +13,12 @@
 #include "strlib.h"
 #include "extern.h"
 
-char client_cert[] = "-----\032\0";
-char private_key[] = "-----\032\0";
-char server_cert[] = "-----BEGIN CERTIFICATE-----\n"
-"MIIEMTCCAxmgAwIBAgIJAM099V95yUdhMA0GCSqGSIb3DQEBBQUAMIGuMQswCQYD\n"
-"VQQGEwJVUzELMAkGA1UECAwCTUkxEjAQBgNVBAcMCUFubiBBcmJvcjEfMB0GA1UE\n"
-"CgwWVW5pdmVyc2l0eSBvZiBNaWNoaWdhbjEkMCIGA1UECwwbUmVhbC10aW1lIFdh\n"
-"dGVyIFN5c3RlbXMgTGFiMRQwEgYDVQQDDAtNYXR0IEJhcnRvczEhMB8GCSqGSIb3\n"
-"DQEJARYSbWRiYXJ0b3NAdW1pY2guZWR1MB4XDTE3MDUyOTAzMjAwOVoXDTE4MDIw\n"
-"ODAzMjAwOVowga4xCzAJBgNVBAYTAlVTMQswCQYDVQQIDAJNSTESMBAGA1UEBwwJ\n"
-"QW5uIEFyYm9yMR8wHQYDVQQKDBZVbml2ZXJzaXR5IG9mIE1pY2hpZ2FuMSQwIgYD\n"
-"VQQLDBtSZWFsLXRpbWUgV2F0ZXIgU3lzdGVtcyBMYWIxFDASBgNVBAMMC01hdHQg\n"
-"QmFydG9zMSEwHwYJKoZIhvcNAQkBFhJtZGJhcnRvc0B1bWljaC5lZHUwggEiMA0G\n"
-"CSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCcdIs+Nt2CwskByoqSUJi9l+H/6y20\n"
-"bZQXDu99v69JXCUGltyss5akBPtbQHWq+hVQSwCXphnYVl2ZsqwKdiz4kuEc/GhT\n"
-"Ng5XqPRWomWC8x3L0xvblvSqYK90tLz0FmzU8zVq6f/OLlTPJZwAhYC8i0mnqbS0\n"
-"KMDvXPA4FfayBhDX9bOUUQos7WoGFQmfT/K/xWlIPmQs2QOFdx6Tp4669JaxnpzZ\n"
-"wSWe7EUidblUbOzCQtKb/XeVQfuW2xdXxQQRr740mY+/w2dHVl0132lypP60nUbk\n"
-"NEVziu4s/C3Lwfb296t4HUfOg460uyzkdDWDZ6NBtSFRNXDhQjeUSs2pAgMBAAGj\n"
-"UDBOMB0GA1UdDgQWBBSJJFBLv/J+ysjgjtHN+FnfTyjIHzAfBgNVHSMEGDAWgBSJ\n"
-"JFBLv/J+ysjgjtHN+FnfTyjIHzAMBgNVHRMEBTADAQH/MA0GCSqGSIb3DQEBBQUA\n"
-"A4IBAQBLGrM/RZKPXyQ2f6ZFA9vugU3KzVgCmV1Z62Io8jOfq8Mhrf0j6s65Yxoh\n"
-"KFqvkrozcG+I44Daz9IZbxU04AsYxhpKN5qO5W2PdS9xOXJWugAciVMrTg510WZd\n"
-"JgYRiYdCk4L72GxvdJ4UKnH1N+t6ix0vAT4e6f/CoLQg6CIhQNjOojR9wz6BkpNn\n"
-"Ra81H5kG3lfajk+o/KPbP4L6CexDnkWiYrkeKPU6SSC7RJ/KqxDHScBfqtmz9OjT\n"
-"xtvtzVID7V7pKFM3j3dON81fLrbDtQtu6XPsBpilfl78rl+hess/tMqnAEHaaqpb\n"
-"54+PblnOoiwRzbySIqKde+lDMCVb\n"
-"-----END CERTIFICATE-----\n"
-"\032\0"; // Make sure to end cert with escape char
-
-// Parameters for SSLSECCFG
-uint8 ssid = 1;
-uint8 cipher_suite = 0; // Use 0 to allow server to decide
-uint8 auth_mode = 1; // 0: No auth, 1: Server, 2: Server/client
-uint8 cert_format = 1; //0: DER, 1: PEM
-
-// Parameters for SSLCFG
-uint8 cid = 1u;
-uint packet_size = 1000u;
-uint max_to = 90u;
-uint def_to = 100u;
-uint tx_to = 50u;
-uint8 ssl_ring_mode = 1u;
-
 // declare variables
 int	   iter = 0;
 uint8  modem_state, lock_acquired = 0u, ready = 0u;
 uint16 uart_string_index = 0u;
 uint32 feed_id;
-char   uart_received_string[1500] = {'\0'};
+char   modem_received_buffer[MODEM_BUFFER_LENGTH] = {'\0'};
 char   request_chunk[CHUNK_SIZE] = {'\0'};
 char*  modem_apn = "epc.tmobile.com";
 
@@ -154,7 +111,7 @@ uint8 at_write_command(char* uart_string, char* expected_response, uint32 uart_t
     
     if( strcmp(expected_response, "") != 0){
         for(i=0;i<interval;i++){
-            char* valid = strstr(uart_received_string, expected_response);
+            char* valid = strstr(modem_received_buffer, expected_response);
             if(valid != NULL){
                 response = 1u;
                 break;
@@ -343,7 +300,7 @@ uint8 modem_get_meid(char *meid) {
         // Expect the UART to contain something like
         // "\r\n#MEID: A10000,32B9F1C0\r\n\r\nOK"
         char *terminator =
-            strextract(uart_received_string, meid, "#MEID: ", "\r\n");
+            strextract(modem_received_buffer, meid, "#MEID: ", "\r\n");
 
         // In the case for modules like CC864-DUAL where "," is in the middle of
         // the MEID, remove the comma
@@ -382,7 +339,7 @@ uint8 modem_check_signal_quality(int *rssi, int *fer) {
         char fer_str[4] = {'\0'};
 
         // Expect the UART to contain something like "+CSQ: 17,99\r\n\r\nOK"
-        char *comma = strextract(uart_received_string, rssi_str, "+CSQ: ", ",");
+        char *comma = strextract(modem_received_buffer, rssi_str, "+CSQ: ", ",");
         char *terminator = strextract(comma, fer_str, ",", "\r\n");
 
         *rssi = atoi(rssi_str);  // Containing "17"
@@ -400,7 +357,7 @@ int modem_get_socket_status() {
         char status_str[5] = {'\0'};
 
         // Expect the UART to contain something like "\r\n#SS: 1,0\r\n"
-        strextract(uart_received_string, status_str, "#SS: ", "\r\n");
+        strextract(modem_received_buffer, status_str, "#SS: ", "\r\n");
 
         // status_str should contain something like "1,0"
         // So increment ptr by 2;
@@ -437,7 +394,7 @@ uint8 modem_pdp_context_toggle(uint8 activate_pdp) {
     // Send AT read command to determine if context is already enabled
     if (at_write_command("AT#SGACT?\r", "OK", 1000u)) {
         // Extract current pdp state into pdp_state
-        strextract(uart_received_string, pdp_state, "SGACT: 1,", "\r\n");
+        strextract(modem_received_buffer, pdp_state, "SGACT: 1,", "\r\n");
 
         // If current state matches desired state, do nothing
         if (atoi(pdp_state) != activate_pdp) {
@@ -452,206 +409,7 @@ uint8 modem_pdp_context_toggle(uint8 activate_pdp) {
     }
 
     return 0u;
-}
-
-uint8 modem_ssl_toggle(int enable_ssl) {
-    char cmd[20];
-    char ssl_state[5] = {'\0'};
-
-    // Send AT read command to determine if SSL is already enabled
-    if (at_write_command("AT#SSLEN?\r", "OK", 1000u)) {
-        // Extract current ssl state into ssl_state
-        strextract(uart_received_string, ssl_state, "SSLEN: 1,", "\r\n");
-
-        // If current state matches desired state, do nothing
-        if (atoi(ssl_state) != enable_ssl) {
-            // Construct AT command
-            sprintf(cmd, "AT#SSLEN=1,%u\r", enable_ssl);
-
-            // Enable/disable SSL
-            return at_write_command(cmd, "OK", 1000u);
-        } else {
-            return 1u;
-        }
-    }
-
-    return 0u;
-}
-
-uint8 modem_gps_power_toggle(uint8 gps_power_on) {
-    char cmd[20];
-    char gps_power[2] = {'\0'};
-
-    // Send AT read command to determine if GPS is already powered
-    if (at_write_command("AT$GPSP?\r", "OK", 1000u)) {
-        // Extract current gps state into gps_power
-        strextract(uart_received_string, gps_power, "GPSP: ", "\r\n");
-
-        // If current state matches desired state, do nothing
-        if (atoi(gps_power) != gps_power_on) {
-            // Construct AT command
-            sprintf(cmd, "AT$GPSP=%u\r", gps_power_on);
-
-            // Enable/disable GPS power
-            return at_write_command(cmd, "OK", 5000u);
-        } else {
-            return 1u;
-        }
-    }
-
-    return 0u;
-}
-
-uint8 modem_get_gps_position(float *lat, float *lon, float *hdop, 
-              float *altitude, uint8 *gps_fix, float *cog, 
-              float *spkm, float *spkn, uint8 *nsat, uint8 min_satellites, uint8 max_tries){
-    uint8 status;
-    int gps_iter;
-    
-        for(gps_iter=0; gps_iter < max_tries; gps_iter++){
-            status = at_write_command("AT$GPSACP\r", "OK", 10000u);
-            if (status){
-                gps_parse(uart_received_string, lat, lon, hdop, altitude, gps_fix, cog, spkm, spkn, nsat);
-                clear_str(uart_received_string);
-                if (*nsat >= min_satellites){
-                    return 1u;
-                }
-                CyDelay(5000u);
-            }
-        }
-    return 0u;
-}
-
-uint8 run_gps_routine(int *gps_trigger, float *lat, float *lon, float *hdop, 
-              float *altitude, uint8 *gps_fix, float *cog, 
-              float *spkm, float *spkn, uint8 *nsat, uint8 min_satellites, uint8 max_tries){
-
-    uint8 status;
-    // Make sure you start with power to GPS off
-    modem_gps_power_toggle(0u);    
-    // Unlock GPS
-    status = at_write_command("AT$GPSLOCK=0\r", "OK", 1000u);
-    if (!status){ return 0u;}
-    CyDelay(100u);
-    // Set antenna path to GPS
-    status = at_write_command("AT$GPSAT=1\r", "OK", 1000u);
-    if (!status){ return 0u;}
-    // Turn on power to GPS
-    modem_gps_power_toggle(1u);
-    CyDelay(10000u);
-    // Get GPS Position
-    status = modem_get_gps_position(lat, lon, hdop, 
-                                    altitude, gps_fix, cog, 
-                                    spkm, spkn, nsat, min_satellites, max_tries);
-    // Turn off power to GPS
-    modem_gps_power_toggle(0u);
-    // Reset GPS Settings
-    at_write_command("AT$GPSRST\r", "OK", 1000u);
-    // For now, always shut gps trigger off
-    (*gps_trigger) = 0u;
-    return status;
-}
-
-uint8 zip_gps(char *labels[], float readings[], uint8 *array_ix,
-              int *gps_trigger, uint8 min_satellites, uint8 max_tries,
-              uint8 max_size) {
-    // Ensure we don't access nonexistent array index
-    uint8 nvars = 10;
-    if (*array_ix + nvars >= max_size) {
-        return *array_ix;
-    }
-
-    float lat = -9999;
-    float lon = -9999;
-    float hdop = -9999;
-    float altitude = -9999;
-    uint8 gps_fix = 0u;
-    float cog = -9999;
-    float spkm = -9999;
-    float spkn = -9999;
-    uint8 nsat = 0u;
-
-    // with the begin/end paradigm, end must always be `one past the end`
-    char **begins = labels + *array_ix;
-    char **ends = begins + nvars;
-    zips(begins, ends, "gps_latitude", "gps_longitude", "gps_hdop",
-         "gps_altitude", "gps_fix", "gps_cog", "gps_spkm", "gps_spkn",
-         "gps_nsat", "gps_trigger");
-
-    run_gps_routine(gps_trigger, &lat, &lon, &hdop, &altitude, &gps_fix, &cog,
-                    &spkm, &spkn, &nsat, min_satellites, max_tries);
-
-    // with the begin/end paradigm, end must always be `one past the end`
-    float *beginf = readings + *array_ix;
-    float *endf = beginf + nvars;
-    zipf(beginf, endf, lat, lon, hdop, altitude, (float) gps_fix, cog, spkm,
-         spkn, (float) nsat, *gps_trigger);
-
-    (*array_ix) += nvars;
-
-    return *array_ix;
-}
-
-uint8 modem_ssl_sec_data(uint8 ssid, uint8 action, uint8 datatype, 
-                         char *cert, char *output_str){
-    char at_command[100] = {'\0'};
-    int certsize;
-    // Construct common portion of AT command
-    sprintf(at_command, "AT#SSLSECDATA=%u,%u,%u", ssid, action, datatype);
-    // Delete mode
-    if (action == 0u){
-        sprintf(at_command, "%s\r", at_command);
-        if (at_write_command(at_command,"OK",1000u)){
-            return 1u;
-        }
-    }
-    // Read mode
-    else if (action == 2u){
-        sprintf(at_command, "%s\r", at_command);
-        if (at_write_command(at_command,"OK",1000u)){
-            strextract(uart_received_string, output_str, "SSLSECDATA: ", "\r\n");
-            return 1u;
-        }
-    }
-    // Write mode
-    else if (action == 1u){
-        // Why is this showing 4?
-        //certsize = sizeof(cert);
-        certsize = strlen(cert) - 1;
-        sprintf(at_command, "%s,%d\r", at_command, certsize);
-        if (at_write_command(at_command,">",1000u)){
-            uart_string_reset();
-		    if(at_write_command(cert,"OK",10000u)){
-                return 1u;
-            }
-        }
-    }
-    return 0u;
-}
-                        
-uint8 modem_ssl_sec_config(uint8 ssid, uint8 cipher_suite, uint8 auth_mode,
-                           uint8 cert_format){
-    char at_command[100] = {'\0'};
-    // Construct AT command
-    sprintf(at_command, "AT#SSLSECCFG=%u,%u,%u,%u\r", ssid, cipher_suite,
-            auth_mode, cert_format);
-    if (at_write_command(at_command,"OK",1000u)){
-            return 1u;
-        }
-    return 0u;    
-}
-                        
-uint8 modem_ssl_config(uint8 ssid, uint8 cid, int packet_size,
-                           int max_to, int def_to, int tx_to, uint8 ssl_ring_mode){
-    char at_command[100] = {'\0'};
-    // Construct AT command
-    sprintf(at_command, "AT#SSLCFG=%u,%u,%u,%u,%u,%u,%u\r", ssid, cid,
-            packet_size, max_to, def_to, tx_to, ssl_ring_mode);
-    if (at_write_command(at_command,"OK",1000u)){
-            return 1u;
-        }
-    return 0u;    
-}                        
+}                    
 
 uint8 modem_socket_dial(char *socket_dial_str, char* endpoint, int port, 
                         int construct_new, int ssl_enabled){
@@ -731,12 +489,14 @@ int send_chunked_request(char* send_str, char *chunk, int chunk_len, char *send_
         a = send_str + i*chunk_len;
         b = send_str + (i+1)*chunk_len;
         if (b >= str_end){
+            // TODO: Reimplement as strncat to avoid unnecessary copying
             strncpy(chunk, a, str_end - a);
             sprintf(chunk, "%s%s", chunk, term_char);
             status = at_write_command(chunk, ring_cmd, 10000u);
             uart_string_reset();
             return status;
         }
+        // TODO: Reimplement as strncat to avoid unnecessary copying
         strncpy(chunk, a, b - a);
         sprintf(chunk, "%s%s", chunk, term_char);
         status = at_write_command(chunk, "OK", 10000u);
@@ -746,7 +506,213 @@ int send_chunked_request(char* send_str, char *chunk, int chunk_len, char *send_
         }        
     }
     return status;
-}                            
+}
+
+int read_response(char message[], char *recv_cmd, char *ring_cmd, uint8 get_response, 
+                  int max_loops, int max_message_size){
+
+    char *chunked_header = "Transfer-Encoding: chunked";
+    char *fixed_length_header = "Content-Length: ";
+    int message_free_space = max_message_size - strlen(message);
+    int chunk_remainder = 0;
+    int buffer_start = 1;
+    int chunk_start = 1;
+    int chunk_size = 0;
+    int status = 0;
+    char *chunked = NULL;
+    char *fixed_length = NULL;
+    char *error = NULL;
+    char *message_end = NULL;
+    char *ending_buffer = NULL;
+    char *response_start = NULL;
+    char length_string[CHUNK_STRING_LENGTH] = {'\0'};
+    char status_code[5] = {"\0"};
+    int iter;
+    
+    char *a = NULL;
+    char *b = NULL;
+
+    // Download the first buffer of data
+    uart_string_reset();
+    int data_pending = at_write_command(recv_cmd, ring_cmd, 10000u);
+    
+    // Check the HTTP response for valid status (200 or 204)
+    parse_http_status(modem_received_buffer, (char*) NULL, status_code, (char*) NULL);
+
+    // If response is not desired, report whether request was successful
+    if ( !get_response){
+        if( status_code[0] == '2' ){
+            return 1u;
+        }
+        return 0u;
+    }
+    // Determine the transfer encoding
+    fixed_length = strstr(modem_received_buffer, fixed_length_header);
+    chunked = strstr(modem_received_buffer, chunked_header);
+
+    // If neither (or both) transfer encodings are included, abort
+    // TODO: This may not always be the case
+    if ((!fixed_length && !chunked) || (fixed_length && chunked)){
+        return 0u;
+    }
+
+    // If fixed length, get the content length
+    if (fixed_length){
+        a = fixed_length + strlen(fixed_length_header);
+        b = strstr(a, "\r\n");
+        if (!b) {return 0u;}
+        // Buffer overflow check
+        if ( (b - a) >= CHUNK_STRING_LENGTH){
+            return 0u;
+        }
+        strncpy(length_string, a, b-a);
+        chunk_size = (int)strtol(length_string, NULL, 10);
+        memset(length_string, '\0', sizeof(length_string));
+        chunk_remainder = chunk_size;
+    }
+
+    // InfluxDB sends chunked data, so check again to see if there is an
+    // unsolicited message indicated a suspended connection
+    //
+    // If so, then the query results are stored in the buffer, so issue
+    // AT#SRECV to read the buffer
+    if (data_pending){
+        // Throw out headers for now
+        uart_string_reset();
+        // Get the next modem buffer
+        status = at_write_command(recv_cmd, "OK", 10000u);
+        // Set the starting pointer to the modem buffer start
+        response_start = modem_received_buffer;
+    }
+    // If no data pending, set the starting pointer
+    else{
+        // If chunked, set the starting pointer right before the first chunk size
+        // If fixed, set the starting pointer at the start of the message body
+            response_start = strstr(modem_received_buffer, "\r\n\r\n");
+    }
+
+    // Make sure response start is not a null pointer
+    if (!response_start){
+        return 0u;
+    }
+    // Set the start and end pointers to the beginning of the message
+    a = response_start;
+    b = response_start;
+
+    // Iterate until all content is downloaded
+    for (iter=0; iter < max_loops; iter++){
+        // Check for CMEE error
+        error = strstr(response_start, "CMEE ERROR");
+        // Get pointer to end of Telit buffer
+        message_end = strstr(response_start, "\r\n\r\nOK\r\n");
+        // Check if this buffer contains the terminating chunk and get pointer
+        ending_buffer = strstr(response_start, "\r\n0\r\n");
+
+        // Break if CMEE error or message truncated
+        if (error || !message_end){
+            return 0u;
+        }
+
+        // Move the beginning pointer to the start of the next line
+        a = strstr(b, "\r\n") + strlen("\r\n");
+        if (!a) {return 0u;}
+        // If we are at the beginning of the Telit buffer, skip another line
+        if (buffer_start){
+            a = strstr(a, "\r\n") + strlen("\r\n");
+            if (!a) {return 0u;}
+            // Reset buffer start flag
+            buffer_start = 0;
+        }
+        // Move the end pointer to the end of the line
+        b = strstr(a, "\r\n");
+        if (!b) {return 0u;}
+
+        // If we're positioned at the start of a new chunk, get the chunk size
+        if (chunked){
+            if (chunk_start){
+                // Buffer overflow check
+                if ( (b - a) >= CHUNK_STRING_LENGTH){
+                    return 0u;
+                }
+                strncpy(length_string, a, b-a);
+                chunk_size = (int)strtol(length_string, NULL, 16);
+                memset(length_string, '\0', sizeof(length_string));
+                chunk_remainder = chunk_size;
+                // Move the beginning pointer to the start of the next line
+                a = b + strlen("\r\n");
+                if (!a) {return 0u;}
+                // Move the end pointer to the end of the line
+                b = strstr(a, "\r\n");
+                if (!b) {return 0u;}
+                // Reset chunk start flag
+                chunk_start = 0;
+            }
+        }
+
+        // Add the current selection to the message buffer
+        // First, check for buffer overflow
+        message_free_space = max_message_size - strlen(message);
+        if ((b - a) >= message_free_space){
+            strncat(message, a, message_free_space);
+            return 0u;
+        }
+        // Concatenate the latest selection to the message
+        strncat(message, a, b - a);
+        // Compute the remaining bytes in the chunk
+        chunk_remainder -= (b - a);
+
+        // If we've reached the terminal chunk, exit the function
+        if (!chunk_size){
+            // return 1 if status code indicates Success, 2xx
+            if( status_code[0] == '2' ){
+                return 1u;
+            }
+            return 0u;
+        }
+
+        // Check if current Telit buffer is exhausted
+        if (b >= message_end){
+            // Make sure that we haven't gone over the end of the buffer
+            if (b > message_end){
+                return 0u;
+            }
+        // If exhausted, get the next Telit buffer
+        status = at_write_command(recv_cmd, "OK", 10000u);
+        // Set the flag to indicate we're at the start of a Telit buffer
+        buffer_start = 1;
+        // Reset pointers
+        response_start = modem_received_buffer;
+        a = response_start;
+        b = response_start;
+        }
+
+        // Check chunk status
+        if (chunk_remainder <= 0){
+            // Check to see if we've overflowed the end of the chunk
+            if (chunk_remainder < 0){
+                return 0u;
+            }
+            // For chunked responses:
+            // If chunk is exhausted, set flag to indicate we're at the 
+            // start of a new chunk
+            if (chunked){
+                if (chunk_size != 0){
+                    chunk_start = 1;
+                }
+            }
+            // For fixed-length responses:
+            // If chunk is exhausted, exit the function
+            else if (fixed_length){
+                // return 1 if status code indicates Success, 2xx
+                if( status_code[0] == '2' ){
+                    return 1u;
+                }
+                return 0u;
+            }
+        }
+    }
+    return 0u;
+}
                             
 uint8 modem_send_recv(char* send_str, char* response, uint8 get_response, int ssl_enabled)
 {
@@ -769,92 +735,20 @@ uint8 modem_send_recv(char* send_str, char* response, uint8 get_response, int ss
     
     // TODO: Should request_chunk be passed in, or global?
     status = send_chunked_request(send_str, request_chunk, CHUNK_SIZE, send_cmd, ring_cmd, "\032");
+    
     if( status ){
-        // Read HTTP response from the buffer
-        uart_string_reset();
-        uint8 data_pending = at_write_command(recv_cmd,ring_cmd,10000u);
-            
-        // Check the HTTP response for valid status (200 or 204)
-        // Create array, "status_code" to temporarily hold the result
-        char status_code[5] = {"\0"};
-        parse_http_status(uart_received_string, (char*) NULL, status_code, (char*) NULL);
-
-        // TODO: Generalize this for N chunks
-        // TODO: Parse for content length
-        // InfluxDB sends chunked data, so check again to see if there is an
-        // unsolicited message indicated a suspended connection
-        //
-        // If so, then the query results are stored in the buffer, so issue
-        // AT#SRECV to read the buffer
-        if (data_pending == 1) {
-            uart_string_reset();
-            status = at_write_command(recv_cmd,"NO CARRIER",10000u);
-        }
-		if (get_response){
-            strcpy(response, uart_received_string);
-		}
-            
-        // return 1 if status code indicates Success, 2xx
-        if( status_code[0] == '2' ){
-            return 1u;
-        }
+        status = read_response(response, recv_cmd, ring_cmd, get_response, 100, MAX_RECV_LENGTH);
+        return status;
     }
-       
+    
     return 0u;  
 }
 
 void uart_string_reset(){
     // reset uart_received_string to zero
-    memset(&uart_received_string[0],0,sizeof(uart_received_string));
+    memset(&modem_received_buffer[0],0,sizeof(modem_received_buffer));
     uart_string_index = 0;
     Telit_UART_ClearRxBuffer();
-}
-
-uint8 ssl_init(uint8 edit_ssl_sec_config, uint8 edit_ssl_config){
-    int response_code = 0;
-    // Enable SSL
-    if(modem_ssl_toggle(1u)){
-        // Edit ssl security settings (SSLSECCFG) if desired
-        if(edit_ssl_sec_config){
-            response_code = modem_ssl_sec_config(ssid, cipher_suite, 
-                                auth_mode, cert_format);
-            if (!response_code){
-                return 0u;
-            }
-        }
-        // Edit general ssl configuration (SSLCFG) if desired
-        if (edit_ssl_config){
-            response_code = modem_ssl_config(ssid, cid, packet_size, max_to, 
-                             def_to, tx_to, ssl_ring_mode);
-            if (!response_code){
-                return 0u;
-            }                            
-        }       
-        // If authorization enabled, store cert data
-        if (auth_mode){
-            response_code = 0;
-            // Delete existing data
-            response_code += modem_ssl_sec_data(ssid, 0, 1, (char*) NULL, (char*) NULL);
-            // Write certificate to NVM
-            response_code += 2*modem_ssl_sec_data(ssid, 1, 1, server_cert, (char*) NULL);
-            if (response_code != 3){
-                return 0u;
-            }
-            if (auth_mode == 2u){
-                // Delete existing data
-                response_code += 4*modem_ssl_sec_data(ssid, 0, 0, (char*) NULL, (char*) NULL);
-                response_code += 8*modem_ssl_sec_data(ssid, 0, 2, (char*) NULL, (char*) NULL);
-                // Write certificate to NVM
-                response_code += 16*modem_ssl_sec_data(ssid, 1, 0, client_cert, (char*) NULL);
-                response_code += 32*modem_ssl_sec_data(ssid, 1, 2, private_key, (char*) NULL);
-                if (response_code != 63){
-                    return 0u;
-                }
-            }
-        }
-        return 1u;
-    }
-    return 0u;
 }
 
 // this function fires when uart rx receives bytes (when modem is sending bytes)
@@ -865,7 +759,7 @@ CY_ISR(Telit_isr_rx){
         
         // store the char in uart_received_string
         if(rx_char_hold){
-            uart_received_string[uart_string_index] = rx_char_hold;
+            modem_received_buffer[uart_string_index] = rx_char_hold;
             uart_string_index++;
         }
     }
