+++ perfect-cell.cydsn/modem.h
@@ -14,10 +14,10 @@
 #define MODEM_STATE_READY       3
 #define MAX_GET_ATTEMPTS        2
 #define MAX_SEND_ATTEMPTS       1//2
-#define MAX_PACKET_LENGTH       1500
 #define MAX_SEND_LENGTH         5000
-#define MAX_RECV_LENGTH         1500
+#define MAX_RECV_LENGTH         5000
 #define CHUNK_SIZE              800
+#define CHUNK_STRING_LENGTH     10
 
 extern uint8	modem_state;
 extern uint8	lock_acquired;
@@ -50,6 +50,17 @@ void modem_start();
  * @brief Deinitialize modem.
  */
 void modem_stop();
+
+/**
+ * @brief Write AT command to cell module
+ *
+ * @param uart_string AT command to write to buffer
+ * @param expected_response Substring that is expected in the response
+ * @param uart_timeout Timeout in milliseconds
+ *
+ * @return 1u if @p expected response is found; 0u otherwise.
+ */
+uint8 at_write_command(char* uart_string, char* expected_response, uint32 uart_timeout);
 
 /**
  * @brief Establish modem connection with internet.
@@ -198,6 +209,21 @@ uint8 modem_socket_close(int ssl_enabled);
 int send_chunked_request(char* send_str, char *chunk, int chunk_len, char *send_cmd, char *ring_cmd, char *term_char);
 
 /**
+ * @brief Reads a HTTP response with a chunked of fixed-length transfer encoding over multiple modem buffers
+ *
+ * @param message The message buffer to write to
+ * @param recv_cmd The Telit receive command to use: AT#SRECV=1,<BYTES>\r or AT#SSLRECV=1,<BYTES>\r
+ * @param ring_cmd The Telit ring URC to seek: SRING: 1 or SSLRING: 1
+ * @param get_response 1u if the response is to be downloaded to the response buffer; 0u otherwise 
+ * @param max_loops The maximum number of times to loop; used as a safeguard
+ * @param max_message_size The maximum size of the @p message
+ *
+ * @return 1u on success, 0u otherwise.
+ */
+int read_response(char message[], char *recv_cmd, char *ring_cmd, uint8 get_response, 
+                  int max_loops, int max_message_size);
+
+/**
  * @brief TODO
  *
  * @param send_str
@@ -238,178 +264,12 @@ void construct_generic_request(char* send_str, char* body, char* host, char* rou
  */
 uint8 modem_pdp_context_toggle(uint8 activate_pdp);
 
-// SSL functions
-
 /**
- * @brief Toggle SSL Socket.
+ * @brief Reset modem_received_buffer, clear RX buffer, and reset string index
  *
- * @param enable_ssl enable flag:
- * - 0 = Disabled
- * - 1 = Enabled
- *
- * @return  1u on success, 0u otherwise.
+ * @return NULL
  */
-uint8 modem_ssl_toggle(int enable_ssl);
-
-/**
- * @brief Stores the security data (certificate(s) and/or private key) into the module’s NVM
- *
- * @param ssid Must be set to 1. It is the only Secure Socket ID available.
- * @param action The action to be performed:
- * - 0 = deleting
- * - 1 = writing
- * - 2 = reading
- * @param datatype identifies the certificate/key to be stored or read:
- * - 0 = Certificate of the client (module). It is needed when the Server/Client authentication
- *   mode has been configured.
- * - 1 = CA Certificate of the remote server, it is used to authenticate the remote server. It is
- *   needed when <auth_mode> parameter of the #SSLSECCFG command is set to 1 or 2.
- * - 2 = RSA private key of the client (module). It is needed if the Server/Client
- *   authentication mode has been configured.
- * @param cert The certificate to be written to the modem.
- * @param output_str The buffer to write the current state of the modem. ONLY USED WHEN ACTION IS READ MODE.
- *
- * @return 1u on success, 0u otherwise.
- */
-uint8 modem_ssl_sec_data(uint8 ssid, uint8 action, uint8 datatype,
-                         char *cert, char *output_str);
-
-/**
- * @brief configure the communication channel according to the user’s security architecture.
- *
- * @param ssid Must be set to 1. It is the only Secure Socket ID available
- * @param cipher_suite Setting the value 0 all the available cipher suites are proposed to the server.
- * It is responsibility of the remote server to select one of them:
- * - 0 = TLS_RSA_WITH_RC4_128_MD5 + TLS_RSA_WITH_RC4_128_SHA + TLS_RSA_WITH_AES_256_CBC_SHA
- * - 1 = TLS_RSA_WITH_RC4_128_MD5
- * - 2 = TLS_RSA_WITH_RC4_128_SHA
- * - 3 = TLS_RSA_WITH_AES_256_CBC_SHA
- * Warning - the product series HE920 / UE910 V2/ DE910 do not support TLS_RSA_WITH_NULL_SHA and TLS_RSA_WITH_AES_256_CBC_SHA.
- * @param auth_mode Is the authentication mode:
- * - 0 = SSL verify none: no authentication, no security data is needed at all
- * - 1 = Server authentication mode: CA Certificate storage is needed (the most common case)
- * - 2 = Server/Client authentication mode: CA Certificate (server), Certificate (client) and Private Key (client) are needed
- * @param cert_format Is an optional parameter. It selects the format of the certificate to be stored via #SSLSECDATA command.
- * - 0 = DER format
- * - 1 = PEM format
- *
- * @return 1u on success, 0u otherwise.
- */
-uint8 modem_ssl_sec_config(uint8 ssid, uint8 cipher_suite, uint8 auth_mode,
-                           uint8 cert_format);
-
-/**
- * @brief Before opening the SSL socket, several parameters can be configured via this command:
- *
- * @param ssid Must be set to 1. It is the only Secure Socket ID available.
- * @param cid The PDP Context Identifier, it's value must be set to 1.
- * @param packet_size Is the size of the packet used by the SSL/TCP/IP stack for data sending in online mode.
- * Small <pktSize> values introduce a higher communication overhead.
- * @param max_to Is the socket inactivity timeout. In online mode: if there’s no data exchange within
- * this timeout period the connection is closed. Increment it if it is needed a longer idle time period.
- * @param def_to Timeout value used as default value by other SSL commands whenever their
- * Timeout parameters are not set.
- * @param tx_to Is the time period after which data is sent even if <pktSize> is not reached (only in online mode).
- * The parameter value must be tuned with user’s application requirements. Small <txTo> values introduce a higher communication overhead.
- * @param ssl_ring_mode Is the presentation mode of the SSLSRING unsolicited indication, which
- * informs the user about new incoming data that can be read in command mode. It can be
- * disabled using value 0.
- *
- * @return 1u on success, 0u otherwise.
- */
-uint8 modem_ssl_config(uint8 ssid, uint8 cid, int packet_size,
-                           int max_to, int def_to, int tx_to, uint8 ssl_ring_mode);
-
-/**
- * @brief Initialize SSL features.
- *
- * @param edit_ssl_sec_config Enable flag: edit SSL security settings (SSLSECCFG) if desired.
- * @param edit_ssl_config Enable flag: edit general SSL configuration (SSLCFG) if desired.
- *
- * @return 1u on success, 0u otherwise.
- */
-uint8 ssl_init(uint8 edit_ssl_sec_config, uint8 edit_ssl_config);
-
-/**
- * @brief Toggle power to the GPS module.
- *
- * @param gps_power_on Enable flag:
- * - 0 = Power off
- * - 1 = Power on
- *
- * @return 1u on success, 0u otherwise.
- */
-uint8 modem_gps_power_toggle(uint8 gps_power_on);
-
-/**
- * @brief Retrieves the current GPS position by attempting to connect to the GPS
- * network.
- *
- * @param lat Buffer to store the Latitude.
- * @param lon Buffer to store the Longitude.
- * @param hdop Buffer to store the Horizontal Diluition of Precision.
- * @param altitude Buffer to store the Altitude: mean-sea-level (geoid).
- * @param gps_fix Buffer to store the flag:
- * - 0 = Invalid Fix
- * - 2 = 2D fix
- * - 3 = 3D fix
- * @param cog Buffer to store the Course over ground.
- * @param spkm Buffer to store the Speed over ground (km/hr).
- * @param spkn Buffer to store the Speed over ground (knots).
- * @param nsat Buffer to store the number of satellites in use [0..12]
- * @param min_satellites Minimum number of satellits the module
- * should connect to.
- * @param max_tries Maximum number of attempts.
- *
- * @return 1u if the gps module was able to connect to the minimum number of satellites,
- * 0u otherwise.
- */
-uint8 modem_get_gps_position(float *lat, float *lon, float *hdop,
-              float *altitude, uint8 *gps_fix, float *cog,
-              float *spkm, float *spkn, uint8 *nsat, uint8 min_satellites, uint8 max_tries);
-
-/**
- * @brief Runs GPS routine to retrieve GPS position.
- *
- * @param gps_trigger gps_trigger flag.
- * @param lat Buffer to store the Latitude.
- * @param lon Buffer to store the Longitude.
- * @param hdop Buffer to store the Horizontal Diluition of Precision.
- * @param altitude Buffer to store the Altitude: mean-sea-level (geoid).
- * @param gps_fix Buffer to store the flag:
- * - 0 = Invalid Fix
- * - 2 = 2D fix
- * - 3 = 3D fix
- * @param cog Buffer to store the Course over ground.
- * @param spkm Buffer to store the Speed over ground (km/hr).
- * @param spkn Buffer to store the Speed over ground (knots).
- * @param nsat Buffer to store the number of satellites in use [0..12]
- * @param min_satellites Minimum number of satellits the module
- * should connect to.
- * @param max_tries Maximum number of attempts.
- *
- * @return 1u if the gps module was able to connect to the minimum number of satellites,
- * 0u otherwise.
- */
-uint8 run_gps_routine(int *gps_trigger, float *lat, float *lon, float *hdop,
-              float *altitude, uint8 *gps_fix, float *cog,
-              float *spkm, float *spkn, uint8 *nsat, uint8 min_satellites, uint8 max_tries);
-
-/**
- * @brief Takes a string array of labels and float array of values and pairs the corresponding values.
- *
- * @param labels[] Array to store the labels.
- * @param readings[] Array to store the corresponding values.
- * @param array_ix  Current array index.
- * @param gps_trigger GPS trigger.
- * @param min_satellites Minimum number of satellits the module
- * should connect to.
- * @param max_tries Maximum number of attempts.
- * @param max_size Max size of the labels and readings arrays.
- *
- * @return (*array_ix) + 10
- */
-uint8 zip_gps(char *labels[], float readings[], uint8 *array_ix, int *gps_trigger, uint8 min_satellites, uint8 max_tries, uint8 max_size);
+void uart_string_reset();
 
 #endif
 //[] END OF FILE
